reading/writing to files
Trying to solve two different problems.
How to keep track of data that has already been extracted from the api.
The modification date seems to be a good candidate to check against since
it is a unique time stamp.

Secondly, how to prepare and present the data so that all updates
are connected with a single user.