Organizing data into a dictionary
Experimented with turning the recovered data into
a dictionary. Each user would be a dictionary with
lists of the companies, urls, and times.

Think I will scrap this idea for now.
Will work on writing all the data to a csv
and database file and will focus on organization/data analysis
afterwards.