Reverse lists
The API returns the data in order from most recent to least recent.
However, we want the most recent updates to be appended to the end of the CSV.

Prior to this, if it updated with 5 new updates at a time, the csv would look like:
5
4
3
2
1
10
9
8
7
6
15
14
13
12
11

Now it should be like this:
1
2
3
4
5
6
7
8
9
10
etc.